feat(ceremony verification): implemented ceremony finalization verification and test fix
Implemented a function that given a ceremony prefix will verify whether the finalization was
successful: all artifacts are uploaded and are valid. These are compared with local copies that are
generated on the fly. Also fixed failing prod tests. Ceremony verification is now a Hardhat task as
it requires smart contract interaction.